requirements.txt
torch=1.11.0
torchvision
matplotlib
Pillow
absl-py
